refactor: refactor the chat service
